microsite-
I put together this microsite just for kicks and to show off my web skills. 🚀 It's all about "The 48 Laws of Power" a controversial book. Each slide is like a mini adventure into one of the book's chapters. I wanted to make it fun and easy to explore, so you'll find the chapter names and some cool context for each law. 📖